refactor: merge of dev of 01/02/2025 (#6)
* feat(push_swap): implement stuff

* feat(push_swap): patch the tester

* style: patch editorconfig

* fix(push_swap): patch radix sort to use stack_max and stack_min

* fix(push_swap): remove some useless stuff

* feat(push_swap): try implementing the new algo

* fix(gpm): patch script

* fix(gnl): remove some useless functions

* fix(push_swap): patch sum

* fix(ford): move

* fix(push_swap): almost there

* fix(push_swap): almost there finally

* fix(push_swap): finalize

* fix(push_swap): fix leaks

* fix(push_swap): upgrade errorhandling

* chore(moulinette.yml): remove useless dirs

* feat(mc): creeper?

* fix(mc): better code

* feat(mc): works kinda but laggy as hell

* refactor(mc): move to files and stuff

* feat(mc): commit de con

* fix(mc): miraculously works

* fix(mc): it works lol

* fix(mc): patch sorting direction

* fix(mc): moved indication to hotbar

* feat(mc): add welcome message

* chore(mc): move to root

* feat(mc): implement clap

* test(mc): implement

* chore(mc): remove some warnings

* refactor(mc): change a bunch of stuff

* chore(.gitignore): remove sume useless stuff

* feat(arthur): implement basi stuff

* docs(arthur): implement basic doc

* feat(arthur): implement a bunch of functionnalities

* feat(arthur): implement better logging

* feat(arthur): implement async

- Implement async running
- Patch some file names

* fix(arthur): implement correct displaying of errors

* fix(arthur): remove loose ends

* style(arthur): apply norm to test files

* style(libft): norm atoi

* style(push_swap): norm

* ci(moulinette): patch indentation

* ci(moulinette): add dispatch call

* ci: fix indentation again

* fix(push_swap): patch compilation errors

* fix(pool): patch all exercises

+++ milestone-1-printf/ft_putnbr_base_ssize.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_base_ssize.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
+/*   By: elagouch <airone01@proton.me>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 13:37:33 by elagouch          #+#    #+#             */
-/*   Updated: 2024/11/20 13:33:15 by elagouch         ###   ########.fr       */
+/*   Updated: 2025/01/02 18:17:30 by elagouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,7 +75,7 @@ ssize_t	ft_putnbr_base_ssize_ulong(uintptr_t nbr, char *base)
 		return (0);
 	count = 0;
 	base_len = ft_strlen(base);
-	if (nbr < 0)
+	if ((long)nbr < 0)
 	{
 		count += write(1, "-", 1);
 		count += ft_putnbr_base_ssize_ulong(-nbr, base);
